+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 **Added**
 - Add `--summary-only` flag.
 
+**Changed**
+- Replace `com.jakewharton.diffuse.io.Size` with `me.saket.bytesize.ByteSize` in the APIs.
+
 **Fixed**
 - Significantly improve `.jar` diff performance.